sampleSize is now double

+++ src/main/java/cubemanager/queryoptimizer/selectivityestimation/SelectivityEstimatorFactory.java
@@ -4,7 +4,7 @@ import cubemanager.cubebase.CubeBase;
 
 public class SelectivityEstimatorFactory {
 
-	public static ISelectivityEstimator create(String method, CubeBase cubeBase, int sampleSize) {
+	public static ISelectivityEstimator create(String method, CubeBase cubeBase, double sampleSize) {
 		switch (method.toUpperCase()) {
 			case "FULL_TABLE_SCAN":
 				return new FullTableScanEstimator(cubeBase);
